requirements.txt
fastapi
uvicorn
pydantic
pandas
langchain
langgraph
openai
python-dotenv
langchain_openai